+++ kotazk_backend/src/main/java/com/taskmanagement/kotazk/payload/request/task/TaskRequestDto.java
@@ -24,7 +24,7 @@ public class TaskRequestDto {
     Long reporterId;
     Long statusId;
     Long priorityId;
-    Long timeEstimate;
+    Float timeEstimate;
     Long timeTracking;
     Boolean isCompleted;
     String startAt;
