+++ kotazk_backend/src/main/java/com/taskmanagement/kotazk/payload/response/task/TaskResponseDto.java
@@ -33,7 +33,7 @@ public class TaskResponseDto {
     Long reporterId;
     StatusResponseDto status;
     PriorityResponseDto priority;
-    Long timeEstimate;
+    Float timeEstimate;
     Long timeTracking;
     Boolean isCompleted;
     Timestamp startAt;
